+++ NEWS.md
@@ -1,6 +1,7 @@
 ## version 1.1.0
 ### Features
 - `check_setup()` function
+- `diagnostics()` function
 
 ## version 1.0.3
 Compatibility update for the rstan 2.31.